+++ src/lpa/__init__.py
@@ -1,7 +1,22 @@
 from ._version import __version__
 from .audit import MutationAuditReport, build_targets, mutation_invariance_audit
-from .config import FeatureConfig, LPAConfig, StudentConfig, TeacherConfig
-from .features import CIFARFeatureExtractor, FeatureViews, coarse48, hog576, row_normalize
+from .config import (
+    FeatureConfig,
+    KMeansFeatureConfig,
+    KMeansTeacherConfig,
+    LinearStudentConfig,
+    LPAConfig,
+    StudentConfig,
+    TeacherConfig,
+)
+from .features import (
+    CIFARFeatureExtractor,
+    FeatureViews,
+    KMeansPatchFeatureExtractor,
+    coarse48,
+    hog576,
+    row_normalize,
+)
 from .federated import (
     RidgeSufficientStatistics,
     aggregate_sufficient_statistics,
@@ -11,22 +26,29 @@ from .federated import (
     packed_symmetric_payload_bytes,
 )
 from .pipeline import LabelPoisoningAntidote
-from .student import LandmarkRidgeStudent
-from .teacher import TrustedKernelTeacher, onehot
+from .student import LandmarkRidgeStudent, LinearRidgeStudent, trusted_row_weights
+from .teacher import BlendedKernelTeacher, TrustedKernelTeacher, onehot
 
 __all__ = [
     "__version__",
     "FeatureConfig",
     "TeacherConfig",
     "StudentConfig",
+    "KMeansFeatureConfig",
+    "KMeansTeacherConfig",
+    "LinearStudentConfig",
     "LPAConfig",
     "CIFARFeatureExtractor",
     "FeatureViews",
+    "KMeansPatchFeatureExtractor",
     "hog576",
     "coarse48",
     "row_normalize",
     "TrustedKernelTeacher",
+    "BlendedKernelTeacher",
     "LandmarkRidgeStudent",
+    "LinearRidgeStudent",
+    "trusted_row_weights",
     "LabelPoisoningAntidote",
     "MutationAuditReport",
     "build_targets",